requirements.txt
flask
python-decouple
tweepy